Reorganize directories and update dialog theme colors
