+++ src/ui/components/configuration/browser/ConfigListItem.vue
