+++ src/ui/components/configuration/browser/MoveFolderTreeNode.vue
@@ -40,7 +40,7 @@ const hasChildren = computed(
 );
 </script>
 
-<style scoped lang="scss">
+<style lang="scss" scoped>
 @use "~/assets/css/colors.scss" as *;
 
 .mftn-row {
